Fixed RerollRules.py formatting
The only functions allowed in the *Rules.py files are one for each
category, with the name of the category. It's an assumption made
elsewhere in the program and it's important that assumption be true. Put
helper functions like singles in Functions.py and call them from the
function named for the category in the *Rules.py files.